+++ src/main/java/com/idukbaduk/itseats/memberaddress/entity/enums/AddressCategory.java
@@ -1,4 +1,4 @@
-package com.idukbaduk.itseats.memberaddress.entity;
+package com.idukbaduk.itseats.memberaddress.entity.enums;
 
 public enum AddressCategory {
     NONE,
